requirements.txt
discord.py==2.3.2
firebase-admin==6.6.0
